Finishing some backsite menu

+++ app/Http/Controllers/Backsite/SpecialistController.php
@@ -32,6 +32,8 @@ class SpecialistController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('specialist_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $specialist = Specialist::orderBy('created_at', 'desc')->get();
 
         // dd($specialist);
@@ -75,6 +77,8 @@ class SpecialistController extends Controller
      */
     public function show(Specialist $specialist)
     {
+        abort_if(Gate::denies('specialist_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         // dd($specialist);
 
         return view('pages.backsite.master-data.specialist.show', compact('specialist'));
@@ -88,6 +92,8 @@ class SpecialistController extends Controller
      */
     public function edit(Specialist $specialist)
     {
+        abort_if(Gate::denies('specialist_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         // dd($specialist);
         
         return view('pages.backsite.master-data.specialist.edit', compact('specialist'));
@@ -121,6 +127,8 @@ class SpecialistController extends Controller
      */
     public function destroy(Specialist $specialist)
     {
+        abort_if(Gate::denies('specilaist_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $specialist->delete();
 
         alert()->success('Success', 'Your specialist has been deleted');
