+++ app/Http/Controllers/Backsite/UserController.php
@@ -46,6 +46,8 @@ class UserController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $user = User::orderBy('created_at', 'desc')->get();
         $type_user = TypeUser::orderBy('name', 'asc')->get();
         $roles = Role::all()->pluck('title','id');
@@ -100,6 +102,8 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        abort_if(Gate::denies('user_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         user->load('role');
 
         return view('pages.backsite.management-access.user.show', compact('user'));
@@ -113,6 +117,8 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        abort_if(Gate::denies('user_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $role = Role::all()->pluck('title','id');
         $type_user = TypeUser::orderBy('name', 'asc')->get();
         $user->load('role');
@@ -151,6 +157,8 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        abort_if(Gate::denies('user_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $user->forceDelete();
         
         alert()->success('Success', 'Your user has been deleted successfully.');
